Small refacto based on change requested.
