+++ src/frontend/web_application/src/modules/encryption/actions/fetchRemoteKeys.js
@@ -1,0 +1,7 @@
+import { requestPublicKeys } from '../../../store/modules/public-key';
+import { tryCatchAxiosAction } from '../../../services/api-client';
+
+export default async function fetchRemoteKeys(dispatch, contactIds) {
+  Promise.all(contactIds.map(contactId =>
+    tryCatchAxiosAction(() => dispatch(requestPublicKeys({ contactId })))));
+}
