+++ src/frontend/web_application/src/modules/encryption/components/LockedMessage/style.scss
@@ -1,6 +1,6 @@
 @import '../../../../styles/common';
 
-.m-encryption-locked_message {
+.m-encryption-locked-message {
   position: relative;
   margin: $co-margin;
   // vertical padding here is meant to avoid margin fusion
