+++ src/frontend/web_application/src/modules/encryption/index.js
@@ -1,5 +1,7 @@
+export * from './actions/getParticipantsKeys.js';
 export { default as CheckDecryption } from './components/CheckDecryption';
 export { default as LockedMessage } from './components/LockedMessage';
 export * from './selectors/message';
 export * from './selectors/publicKey';
 export * from './services/keyring/remoteKeys';
+
